Changed Color Scheme of Site and Added Footer
Changed to a green color scheme for the site, instead of black and white. Also added a footer and populated it with a roadmap, as well as social media links. Also changed the Nav.Brand in the Navbar to a logo image, which now goes to Home instead of a listed Home link.